# Update Benchmark to reflect optimisations