--- a/csharp/safe-efficient-code/benchmark/Program.cs
+++ b/csharp/safe-efficient-code/benchmark/Program.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Running;
 
@@ -36,9 +36,9 @@ namespace benchmark
 
     public struct MyMutableStruct
     {
-        public double X { get; }
-        public double Y { get; }
-        public double Z { get; }
+        public double X { get => x; set => x = value; }
+        public double Y { get => y; set => y = value; }
+        public double Z { get => z; set => z = value; }
 
         private double a;
         private double b;
@@ -48,19 +48,56 @@ namespace benchmark
         private double f;
         private double g;
         private double h;
+        private double z;
+        private double y;
+        private double x;
+
         public MyMutableStruct(double x, double y = 0, double z = 0)
         {
-            X = x;
-            Y = y;
-            Z = z;
-            a = 1;
-            b = 2;
-            c = 3;
-            d = 4;
-            e = 5;
-            f = 6;
-            g = 7;
-            h = 8;
+            this.x = x;
+            this.y = y;
+            this.z = z;
+            this.a = 1;
+            this.b = 2;
+            this.c = 3;
+            this.d = 4;
+            this.e = 5;
+            this.f = 6;
+            this.g = 7;
+            this.h = 8;
+        }
+    }
+    public struct MyMutableStructReadonly
+    {
+        public double X { readonly get => x; set => x = value; }
+        public double Y { readonly get => y; set => y = value; }
+        public double Z { readonly get => z; set => z = value; }
+
+        private double a;
+        private double b;
+        private double c;
+        private double d;
+        private double e;
+        private double f;
+        private double g;
+        private double h;
+        private double z;
+        private double y;
+        private double x;
+
+        public MyMutableStructReadonly(double x, double y = 0, double z = 0)
+        {
+            this.x = x;
+            this.y = y;
+            this.z = z;
+            this.a = 1;
+            this.b = 2;
+            this.c = 3;
+            this.d = 4;
+            this.e = 5;
+            this.f = 6;
+            this.g = 7;
+            this.h = 8;
         }
     }
 
@@ -68,17 +105,18 @@ namespace benchmark
     {
         MyImmutableStruct sStruct = new MyImmutableStruct(1.1, 2.2);
         MyMutableStruct mStruct = new MyMutableStruct(1.1, 2.2);
+        MyMutableStructReadonly mrStruct = new MyMutableStructReadonly(1.1, 2.2);
 
         [Benchmark]
-        public void ImmutableAddByType()
+        public double ImmutableAddByType()
         {
-            add_by_type(sStruct);
+            return add_by_type(sStruct);
         }
 
         [Benchmark]
-        public void ImmutableAddByRefType()
+        public double ImmutableAddByRefType()
         {
-            add_by_reftype(in sStruct);
+            return add_by_reftype(in sStruct);
         }
         public double add_by_type(MyImmutableStruct s)
         {
@@ -90,15 +128,15 @@ namespace benchmark
         }
 
         [Benchmark]
-        public void MutableAddByType()
+        public double MutableAddByType()
         {
-            add_by_type(mStruct);
+            return add_by_type(mStruct);
         }
 
         [Benchmark]
-        public void MutableAddByRefType()
+        public double MutableAddByRefType()
         {
-            add_by_reftype(in mStruct);
+            return add_by_reftype(in mStruct);
         }
 
         public double add_by_type(MyMutableStruct s)
@@ -106,6 +144,27 @@ namespace benchmark
             return s.X + s.Y;
         }
         public double add_by_reftype(in MyMutableStruct s)
+        {
+            return s.X + s.Y;
+        }
+
+        [Benchmark]
+        public double MutableReadOnlyAddByType()
+        {
+            return add_by_type(mrStruct);
+        }
+
+        [Benchmark]
+        public double MutableReadonlyAddByRefType()
+        {
+            return add_by_reftype(in mrStruct);
+        }
+
+        public double add_by_type(MyMutableStructReadonly s)
+        {
+            return s.X + s.Y;
+        }
+        public double add_by_reftype(in MyMutableStructReadonly s)
         {
             return s.X + s.Y;
         }